+++ vaping/plugins/zeromq.py
@@ -20,18 +20,33 @@ class ZeroMQ(vaping.plugins.EmitBase):
             return
 
         self.skip = False
-
         self.ctx = zmq.Context()
+
+        # sanity check config
+        if 'bind' in self.config:
+            if 'connect' in self.config:
+                msg = "bind and connect are mutually exclusive"
+                self.log.error(msg)
+                raise ValueError(msg)
+
+        elif 'connect' not in self.config:
+            self.log.warning("skipping zeromq, missing bind or connect")
+            self.skip = True
+
+    def on_start(self):
+        if self.skip:
+            return
         self.sock = self.ctx.socket(zmq.PUB)
         if 'bind' in self.config:
             self.sock.bind(self.config['bind'])
-        else:
-            self.log.warning("skipping zeromq, missing config")
-            self.skip = True
+        elif 'connect' in self.config:
+            self.sock.connect(self.config['connect'])
 
-
-# TODO option topic
-#        self.topic = 'vaping.fping'
+    def on_stop(self):
+        if self.skip:
+            return
+        if self.sock:
+            self.sock.close()
 
     def emit(self, data):
         if self.skip: