+++ CHANGELOG.md
@@ -4,8 +4,16 @@
 ## [Unreleased]
 ### Added
 - py3 support
+- added on_start() and on_stop() events to plugins
+- zeromq can connect or bind to socket
+
 ### Fixed
+- #2 error if zeromq is missing
+- #3 daemonize closes plugin fds
+
 ### Changed
+- call start() on emit plugins
+
 ### Deprecated
 ### Removed
 ### Security